+++ src/losant/contains-any.test.js
@@ -1,0 +1,20 @@
+const { pipe, map, apply, adjust } = require('ramda');
+
+const testHarness = require('./test-harness');
+const containsAny = require('./contains-any');
+
+const cases = [
+  [[1, 2], [1, 2, 3], true],
+  [['foo'], ['foo', 'bar'], true],
+  [[null], [null, 'foo'], true],
+  [[1], ['foo'], false],
+  [[], ['foo'], false],
+  [['foo'], [], false],
+];
+
+const testHarnessAdapter = pipe(
+  adjust(containsAny, 0),
+  apply(testHarness),
+);
+
+map(testHarnessAdapter)(cases);