+++ src/losant/is-populated-string.test.js
@@ -1,44 +1,18 @@
-const { map, curry } = require('ramda');
-const { isObject } = require('lodash/fp');
+const { map, apply } = require('ramda');
+
+const testHarness = require('./test-harness');
 const isPopulatedString = require('./is-populated-string');
 
-// harness :: Boolean -> * -> Function -> Boolean
-const harness = curry((expected, input) => {
-  return () => {
-    expect(
-      isPopulatedString(input)
-    ).toEqual(expected);
-  };
-});
+const cases = [
+  ['foo', true],
+  ['   ', true],
+  ['', false],
+  [5, false],
+  [false, false],
+  [{ foo: 'foo' }, false],
+  [[ 'foo' ], false],
+];
 
-describe('isPopulatedString', () => {
-
-  describe('true cases', () => {
-    const inputs = [
-      'foo',
-      '   '
-    ];
-    map(
-      (input) => { test(`case: \`${input}\``, harness(true, input)); },
-      inputs
-    );
-  });
-
-  describe('false cases', () => {
-    const inputs = [
-      '',
-      5,
-      false,
-      { foo: 'foo' },
-      [ 'foo' ]
-    ];
-    map(
-      (input) => {
-        const label = isObject(input) ? JSON.stringify(input) : input;
-        test(`case: \`${label}\``, harness(false, input));
-      },
-      inputs
-    );
-  });
-
-});
+map(
+  apply(testHarness(isPopulatedString))
+)(cases);