+++ src/ramda/index.js
@@ -1,3 +1,9 @@
 const r = require('ramda');
 
-module.exports = r.omit([], r);
+module.exports = r.omit([
+  'toString',
+  'toLower',
+  'toUpper',
+  'trim',
+  'split',
+], r);