+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 Below should serve as an "upgrade guide" jumping off point as you are migrating projects to newer versions of this package. Ideally, interface changes will be called out specifically and one should not have to go sifting through git diffs to discern what changed.
 
 - `?.?.?` - *WIP*
+   - Added methods:
+      - `containsAny`
+      - `isNotNil`
+   - The following methods are now pulled from Lodash instead of Ramda:
+      - `toString`
+      - `toLower`
+      - `toUpper`
+      - `trim`
+      - `split`
 - `1.0.9`
    - Added methods:
       - `jsonParseSafe`